Fix some fields of Student and Catalog Course entities
This commit fixes the data type of the 'EnrollmentDate' field of the 'Student' entity and the 'CourseYear' field of the 'CatalogCourse' entity.

Before this commit, the equivalent of these fields in database tables was 'Year'. ModelHelper was not working properly when trying to bind data from database to objects. This is because the datatype 'Year' in the database corresponds to the 'short' data type in C#.

This commit updates the corresponding fields of the related entities. It also adds a binding code for the 'short' data type to ModelHelper.